Fixed the way text fields' constraints are set.
This change ensures that the text fields are always presented on top of the image, and that they remain in their place on top of the image even if the device was rotated. In order to ensure that, the text fields' view-edge constraints (depending on the text field) were replaced with centre-y-view constraints; these are updated depending on the size of the image displayed.
*Image size was also corrected - PHImageManager now fetches the full quality image.